Update MoveImagesForTagging.py
Moved images to tag to a new folder so that processing and tagging can happen at the same time. Previously, I was using the ImagesToProcess folder for tagging when I needed to and processing at different times. Also decreased the number of images to tag at a time to 10. It was taking too long to finish a set of images.