Update MoveImagesForTagging.py
Added better comments. Automated a few parts of the process. Now the zip file with predictions is automatically opened and extracted into the folder of human tags. The files that were copied into the totag folder are deleted, and then the new set of images to predict are moved into the totag folder. This reduces the amount of manual copying and deleting required.