Create CreateVideo.py
Initial code dump to  create a video of the frames from a set of lifting. Currently missing features. Built a list of the frames inside a single set, and displayed those images in order. Need to load in the prediction boxes, add plots of the locations vs time, and then save the results as a video.